Refactor data access and add test project
Refactored data access to use Dapper and SqlConnection instead of Entity Framework, with corresponding updates in workspace settings. Additional methods were added for lock acquisition and release, and the lock timeout setting in DbConnectionSettings was updated. Moreover, a new test project, DistributedLockTest, was introduced, with NUnit being used for the test framework.