Update DataModel and adjust namespaces, refactor SQL operations
Minor refactorings have been applied to namespaces throughout the solution, changing them from the semicolon syntax to the more conventional block syntax. In addition, a significant update has been introduced to the DataModel, including the creation of new classes and enums such as Job, JobStatus, and others. The data access layer has been tweaked as well, including adjustments in SQL operations, particularly the way locks are acquired and released in the DataLockDBContext, and connection string was updated in the appsettings file.